allow starting custom ns_memcached instances. MB-51-something
Custom ns_memcached instances are useful when we need to do something
that ns_memcached on old node doesn't support. Or if we're doing some
data operations and we need more than one instance.

Note that custom ns_memcached's (rightfully) don't have logic of
creating bucket and waiting warmup. The idea is that whoever spawns
it will do it on ready bucket.

Change-Id: I34bb979358fa48e4e32faf0156bf7c808150fd78