wip:MB-5307: use past seqno in replicator's confirm_sent_messages
NOTE:alk: After some thinking I don't think I need impossible
seqno. Just distinct from what's on the wire.

So that we can continue using downstream after using
confirm_sent_messages.

We previously assumed that downstream is not going to be reused after
confirm_sent_messages is called. We used it as part of replicator
shutdown.

But we want it to be part of vbucket filter change operation that
reuses downstream socket. And thus we're using 'impossible' seqno in
order to be able to reuse downstream socket.

Change-Id: I46ee22fe0b7cb9160821d53459b18c3d482e960b