MB-100 Replace randomized logging with special logging category.
I would argue that the only useful values for the trace_dump_invprob
were 1 and infinity. Given that, I replaced all the randomized logging
with a logging to a special logger called ?XDCR_EXTRA_DEBUG. And
xdc_rep_manager adjusts the threshold loglevel for this logger based
on the value of xdcr_enable_extra_debug config parameter.

Change-Id: I5a479b89c1b015d5a543b694622a1f8d4e37a7d7